Port the rest of the implementation of PlayMusic
Port the rest of the implementation fo PlayMusic.

Note: This is a WIP; I still haven't figured out how to build webmidi so I can't test it at all, and the Thread.Sleeps don't have an easy TS/JS equivalent that I can see so I've got a placeholder in place there pending refactoring of the code.